feat: add Anthropic bash tool renderer with shared bash display components
Extract common bash UI into BashDisplay components for both CLI and web,
then create dedicated bash_anthropic renderers that normalize content-block
output into the shared display.
